requirement.txt
APScheduler==3.6.1
asgiref==3.2.3
asn1crypto==1.2.0
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
blinker==1.4
brotlipy==0.7.0
certifi==2019.9.11
cffi==1.13.2
channels==2.3.1
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.3.1
cx-Oracle==7.2.3
daphne==2.3.0
Django==2.2.3
django-cors-headers==3.1.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
kaitaistruct==0.8
ldap3==2.5.2
mitmproxy==4.0.4
passlib==1.7.1
Pillow==6.2.1
pyasn1==0.4.7
pycparser==2.19
pyDes==2.0.1
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
pytz==2019.3
redis==3.3.11
requests==2.22.0
ruamel.yaml==0.15.100
six==1.12.0
sortedcontainers==2.0.5
sqlparse==0.3.0
tornado==5.1.1
Twisted==19.7.0
txaio==18.8.1
tzlocal==2.0.0
urllib3==1.25.6
urwid==2.0.1
wsproto==0.11.0
xlrd==1.2.0
zope.interface==4.6.0
service_identity
pymysql
mysqlclient==1.3.14
python-jenkins==1.6.0
easycython
pyinstaller
django-debug-toolbar
celery
channels_redis==2.4.2
python-gitlab==2.4.0
pymongo==3.10.1